test: add more fail tests for ddl ingest mode
